update the raspbian recipe enough to insert a custom_hook + update the recipe runner to recusively include recipes

+++ tools/util/recipe_runner.sh
@@ -1,23 +1,38 @@
 #!/bin/bash
-export CFGDIR=$TOPDIR"/tools/cfg"
-source $TOPDIR/bin/"activate"
+RECIPE_BUILD_LEVEL=${RECIPE_BUILD_LEVEL:-0}
 
-rm -rf $BUILDDIR/*
-mkdir -p $BUILDDIR/output
-source $UTILDIR/parameters.sh
-
-echo $BUILDDIR/raw_recipe.sh
-
-$UTILDIR/recipe_aide.py -r "$@" -o $BUILDDIR/raw_recipe.sh
-if [ $? -ne 0 ]; then
-    echo "failed processing recipe"
+if [ $RECIPE_BUILD_LEVEL -eq 0 ]; then
+    export CFGDIR=$TOPDIR"/tools/cfg"
+    source $TOPDIR/bin/"activate"
+    rm -rf $BUILDDIR/*
+    mkdir -p $BUILDDIR/output
+    source $UTILDIR/parameters.sh
 fi
 
-# run recipe
-source $BUILDDIR/raw_recipe.sh
+RAW_RECIPE_PATH=$BUILDDIR/raw_recipe_$RANDOM.sh
+echo "$@ ==maps=to==> $RAW_RECIPE_PATH"
+$UTILDIR/recipe_aide.py -r "$@" -o $RAW_RECIPE_PATH
+if [ $? -ne 0 ]; then
+    echo "failed processing recipe $@"
+    exit 1
+fi
+source $RAW_RECIPE_PATH
+
+for RINC in ${RECIPE_INCLUDES[*]}
+do
+    let RECIPE_BUILD_LEVEL=($RECIPE_BUILD_LEVEL + 1)
+    export RECIPE_BUILD_LEVEL
+    ${TOPDIR}/bin/myenv.sh "${TOPDIR}/recipes/${RINC}.yml"
+    let RECIPE_BUILD_LEVEL=($RECIPE_BUILD_LEVEL - 1)
+done
+
+echo "recursive copy $RECIPE_TEMPLATE/* to $BUILDDIR/"
+cp -R $RECIPE_TEMPLATE/* $BUILDDIR/
+if [ $RECIPE_BUILD_LEVEL -gt 0 ]; then
+    exit 0
+fi
 
 if [ "$RECIPE_EXPAND_TEMPLATE" == "yes" ]; then
-    cp -R $RECIPE_TEMPLATE/* $BUILDDIR/
     for MUSTACHE_FILE in $(find $BUILDDIR -type f | grep ".mustache$")
     do
         $UTILDIR/filler.py -t $MUSTACHE_FILE -p $PARAM_FILE
